feat(sql): supports schemas in read_json, read_csv, read_parquet
