+++ src/daft-sql/Cargo.toml
@@ -12,6 +12,7 @@ daft-functions-json = {path = "../daft-functions-json"}
 daft-logical-plan = {path = "../daft-logical-plan"}
 daft-scan = {path = "../daft-scan"}
 daft-session = {path = "../daft-session"}
+indexmap = {workspace = true}
 once_cell = {workspace = true}
 pyo3 = {workspace = true, optional = true}
 sqlparser = {workspace = true}
