+++ src/daft-sql/src/table_provider/read_csv.rs
@@ -1,3 +1,5 @@
+use std::sync::Arc;
+
 use daft_logical_plan::LogicalPlanBuilder;
 use daft_scan::builder::CsvScanBuilder;
 use sqlparser::ast::TableFunctionArgs;
@@ -9,6 +11,7 @@ use crate::{
     invalid_operation_err,
     modules::config::expr_to_iocfg,
     planner::SQLPlanner,
+    schema::try_parse_schema,
 };
 
 pub(super) struct ReadCsvFunction;
@@ -43,7 +46,11 @@ impl TryFrom<SQLFunctionArguments> for CsvScanBuilder {
         let buffer_size = args.try_get_named("buffer_size")?;
         let file_path_column = args.try_get_named("file_path_column")?;
         let hive_partitioning = args.try_get_named("hive_partitioning")?.unwrap_or(false);
-        let schema = None; // TODO
+        let schema = if let Some(schema_arg) = args.try_get_named("schema")? {
+            Some(Arc::new(try_parse_schema(schema_arg)?))
+        } else {
+            None
+        };
         let schema_hints = None; // TODO
         let io_config = args.get_named("io_config").map(expr_to_iocfg).transpose()?;
 
@@ -79,7 +86,7 @@ impl SQLTableFunction for ReadCsvFunction {
             &[
                 "path",
                 "infer_schema",
-                // "schema",
+                "schema",
                 "has_headers",
                 "delimiter",
                 "double_quote",
