+++ src/daft-sql/src/table_provider/read_json.rs
@@ -1,7 +1,12 @@
+use std::sync::Arc;
+
 use daft_scan::builder::JsonScanBuilder;
 
 use super::{expr_to_iocfg, try_coerce_list, SQLTableFunction};
-use crate::{error::PlannerError, functions::SQLFunctionArguments, invalid_operation_err};
+use crate::{
+    error::PlannerError, functions::SQLFunctionArguments, invalid_operation_err,
+    schema::try_parse_schema,
+};
 
 pub(super) struct ReadJsonFunction;
 
@@ -16,7 +21,7 @@ impl SQLTableFunction for ReadJsonFunction {
             &[
                 "path",
                 "infer_schema",
-                // "schema"
+                "schema",
                 "io_config",
                 "file_path_column",
                 "hive_partitioning",
@@ -53,7 +58,11 @@ impl TryFrom<SQLFunctionArguments> for JsonScanBuilder {
         let buffer_size = args.try_get_named("buffer_size")?;
         let file_path_column = args.try_get_named("file_path_column")?;
         let hive_partitioning = args.try_get_named("hive_partitioning")?.unwrap_or(false);
-        let schema = None; // TODO
+        let schema = if let Some(schema_arg) = args.try_get_named("schema")? {
+            Some(Arc::new(try_parse_schema(schema_arg)?))
+        } else {
+            None
+        };
         let schema_hints = None; // TODO
         let io_config = args.get_named("io_config").map(expr_to_iocfg).transpose()?;
